% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatCNA.R
\name{formatCNA}
\alias{formatCNA}
\title{formatC with custom NA-string}
\description{
  Allows to use a custom NA-string in \code{\link{formatC}} instead of
  the defaults \sQuote{NA}.
}
\usage{
formatCNA(x, naStr = "", ...)
}
