\name{rSquared-methods}
\docType{methods}
\alias{rSquared-methods}
\alias{rSquared,lm-method}
\title{R-squared-related methods}
\description{
  R-squared and related methods for different statistical models.
  Currently only for 'lm' class.
}
\section{Methods}{
\describe{
\item{\code{signature(object = "lm")}}{
  Ordinary \eqn{R^2}{R-squared} for estimated linear models
}
}}
\keyword{methods}
